mal: Make some optional functions optional.
There are some functions that the outer implementation isn't required
to provide.  If those functions aren't available, a self-hosted mal
should still work, but it can't provide those functions itself.
Arrange that this is so by wrapping each of them in a trivial function
so that the undefined function is only detected if the wrapper
function is called.