apl: Make (str "") work correctly
It appears that in APL, ∈⊂'' isn't guaranteed to yield a string.  At
least, I think the recursive definition in the standard boils down to
joining an empty string and an empty mixed array, and the result of
that is implementation-defined.

In GNU APL, it looks if I append another empty string I get the right
answer, so do that.